NDC_local/requirements.txt
addict==2.4.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.2.11
django-cors-headers==3.14.0
django-mysql==3.9.0
djangorestframework==3.15.0
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.5.0
gcloud==0.18.3
googleapis-common-protos==1.63.0
gunicorn==22.0.0
httplib2==0.15.0
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jws==0.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
open3d==0.18.0
packaging==24.0
pandas==2.2.2
pika==1.3.2
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
PyMySQL==1.1.0
pyparsing==3.1.2
pyquaternion==0.9.9
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-jwt==2.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.29.0
requests-toolbelt==0.10.1
retrying==1.3.4
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tenacity==8.3.0
threadpoolctl==3.5.0
torch==2.3.0
tqdm==4.66.4
traitlets==5.14.3
trimesh==4.4.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.3
zipp==3.18.2
